# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Portfolio API",
    "description": "Backend API for portfolio showcase with user management, projects, skills, and analytics",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "hyoukjoo@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "Get all projects",
        "description": "Retrieve a list of portfolio projects with optional filtering",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "type": "string",
            "description": "Filter by project status",
            "enum": ["in-progress", "completed", "planned"]
          },
          {
            "in": "query",
            "name": "featured",
            "type": "boolean",
            "description": "Filter by featured projects"
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Project"
                  }
                },
                "count": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["projects"],
        "summary": "Create a new project",
        "description": "Add a new project to the portfolio",
        "parameters": [
          {
            "in": "body",
            "name": "project",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateProjectRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Project created successfully",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        }
      }
    },
...
    "/contact": {
      "post": {
        "tags": ["contact"],
        "summary": "Submit contact form",
        "description": "Submit a contact form message",
        "parameters": [
          {
            "in": "body",
            "name": "contact",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ContactFormRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact form submitted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "id": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Project": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "title": {
          "type": "string",
          "example": "Portfolio Website"
        },
        "description": {
          "type": "string",
          "example": "A responsive portfolio website built with Flutter"
        },
        "tech_stack": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["Flutter", "Dart", "GitHub Pages"]
        },
        "status": {
          "type": "string",
          "enum": ["in-progress", "completed", "planned"],
          "example": "completed"
        },
        "featured": {
          "type": "boolean",
          "example": true
        },
        "live_url": {
          "type": "string",
          "example": "https://johndoe.github.io/portfolio"
        },
        "github_url": {
          "type": "string",
          "example": "https://github.com/johndoe/portfolio"
        },
        "image_url": {
          "type": "string",
          "example": "https://example.com/project-image.jpg"
        },
        "start_date": {
          "type": "string",
          "format": "date-time"
        },
        "end_date": {
          "type": "string",
          "format": "date-time"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CreateProjectRequest": {
      "type": "object",
      "required": ["title", "description", "status"],
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "tech_stack": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "status": {
          "type": "string",
          "enum": ["in-progress", "completed", "planned"]
        },
        "featured": {
          "type": "boolean"
        },
        "live_url": {
          "type": "string"
        },
        "github_url": {
          "type": "string"
        },
        "image_url": {
          "type": "string"
        },
        "start_date": {
          "type": "string",
          "format": "date-time"
        },
        "end_date": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "ContactFormRequest": {
      "type": "object",
      "required": ["name", "email", "subject", "message"],
      "properties": {
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "subject": {
          "type": "string",
          "example": "Project Inquiry"
        },
        "message": {
          "type": "string",
          "example": "I would like to discuss a project opportunity"
        }
      }
    },
...
  }
}